requirements.txt
flask==2.3.1
requests==2.29.0
python-dotenv==1.0.0
pandas==2.0.3
jupyter==1.0.0
gunicorn==21.2.0